\name{laercio-package}
\alias{laercio-package}
\alias{laercio}
\docType{package}
\title{
Functions for Means comparison.
}
\description{
The package contains tests to compare means.
}
\details{
\tabular{ll}{
Package: \tab laercio\cr
Type: \tab Package\cr
Version: \tab 0.9-1\cr
Date: \tab 2008-06-10\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Laercio Junio da Silva

Maintainer: Laercio Junio da Silva \email{laerciojunio@yahoo.com.br}
}
\keyword{ package }
}