\name{laercio-package}
\alias{laercio-package}
\alias{laercio}
\docType{package}
\title{
Duncan test, Tukey test and Scott-Knott test.
}
\description{
The package contains functions to compare and group means.
}
\details{
\tabular{ll}{
Package: \tab laercio\cr
Type: \tab Package\cr
Version: \tab 0.9-4\cr
Date: \tab 2008-07-15\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Laercio Junio da Silva

Maintainer: Laercio Junio da Silva \email{laerciojunio@yahoo.com.br}
}
\keyword{ package }
}